team34/src/main/ai-model/requirements.txt
Pillow
keras
numpy
flask
gevent
tensorflow